github-portfolio-analyzer/backend/requirements.txt
# Backend requirements for GitHub Recruiter Analyzer

fastapi
uvicorn
requests
python-dotenv
